Makes properties nullable to prevent errors
Addresses potential null reference exceptions by making several properties nullable, especially in data models and DTOs used during import processes.

Adds a null check for supplier in the katalog controller to prevent errors when a supplier is not assigned to an asset.

Returns an empty array in proxy controller when no asset id is set